feat: 1.调整新增 EventHookMgr 用以管理 Pre-fire hook 及 Post-fire hook 2.增加 LLBC_CFG_CORE_ENABLE_EVENT_HOOK 用以开启或关闭编译 event hook 相关代码 3.event hook 相关内容空间惰式创建 #396
